Поправлен вызов моделей. Создан pipeline и конфиг для него, где можно писать какую модель хочется запустить. В соответствии с этим на мой взгляд повышается масштабируемость сервиса и код с ручкой можно в принципе не трогать, т.к. формирование реков будет происходить в зависимости от того, какую модель определи при запуске.
