+++ service/config/main-pipeline.cfg.yml
@@ -1,0 +1,3 @@
+type_model:
+  user_knn: True
+  matrix_factorization: False
