--- service/utils/run_reco_mf.py
@@ -1,3 +1,0 @@
-from service.utils.reco_mf import RecommendMF
-
-mf_model = RecommendMF()
